anvil: Split up shell module
