--- anvil/src/shell.rs
@@ -1,1410 +1,0 @@
-use std::{cell::RefCell, time::Duration};
-
-use smithay::{
-    backend::renderer::{
-        element::{surface::WaylandSurfaceRenderElement, AsRenderElements},
-        utils::on_commit_buffer_handler,
-        ImportAll, Renderer,
-    },
-    desktop::{
-        find_popup_root_surface, layer_map_for_output,
-        space::SpaceElement,
-        utils::{
-            send_frames_surface_tree, take_presentation_feedback_surface_tree, under_from_surface_tree,
-            with_surfaces_surface_tree, OutputPresentationFeedback,
-        },
-        LayerSurface, PopupKeyboardGrab, PopupKind, PopupManager, PopupPointerGrab, PopupUngrabStrategy,
-        Space, Window, WindowSurfaceType,
-    },
-    input::{
-        pointer::{
-            AxisFrame, ButtonEvent, Focus, GrabStartData as PointerGrabStartData, MotionEvent, PointerGrab,
-            PointerInnerHandle,
-        },
-        Seat,
-    },
-    output::Output,
-    reexports::{
-        wayland_protocols::{
-            wp::presentation_time::server::wp_presentation_feedback, xdg::shell::server::xdg_toplevel,
-        },
-        wayland_server::{
-            protocol::{wl_buffer::WlBuffer, wl_output, wl_seat, wl_surface::WlSurface},
-            Resource,
-        },
-    },
-    space_elements,
-    utils::{IsAlive, Logical, Point, Rectangle, Serial, Size, SERIAL_COUNTER},
-    wayland::{
-        buffer::BufferHandler,
-        compositor::{
-            get_parent, is_sync_subsurface, with_states, with_surface_tree_upward, CompositorHandler,
-            CompositorState, SurfaceData as WlSurfaceData, TraversalAction,
-        },
-        seat::WaylandFocus,
-        shell::{
-            wlr_layer::{
-                Layer, LayerSurface as WlrLayerSurface, LayerSurfaceData, WlrLayerShellHandler,
-                WlrLayerShellState,
-            },
-            xdg::{
-                Configure, PopupSurface, PositionerState, SurfaceCachedState, ToplevelSurface,
-                XdgPopupSurfaceData, XdgShellHandler, XdgShellState, XdgToplevelSurfaceData,
-            },
-        },
-    },
-    xwayland::{
-        xwm::{ResizeEdge as X11ResizeEdge, XwmId},
-        X11Surface, XwmHandler, X11WM,
-    },
-};
-
-use crate::{
-    focus::FocusTarget,
-    state::{AnvilState, Backend},
-    CalloopData,
-};
-
-#[cfg(not(feature = "xwayland"))]
-space_elements! {
-    #[derive(Debug, Clone, PartialEq)]
-    pub WindowElement;
-    Wayland=Window,
-}
-
-#[cfg(feature = "xwayland")]
-space_elements! {
-    #[derive(Debug, Clone, PartialEq)]
-    pub WindowElement;
-    Wayland=Window,
-    X11=X11Surface,
-}
-
-impl WindowElement {
-    pub fn surface_under(
-        &self,
-        location: Point<f64, Logical>,
-        window_type: WindowSurfaceType,
-    ) -> Option<(WlSurface, Point<i32, Logical>)> {
-        match self {
-            WindowElement::Wayland(w) => w.surface_under(location, window_type),
-            #[cfg(feature = "xwayland")]
-            WindowElement::X11(w) => w
-                .wl_surface()
-                .and_then(|s| under_from_surface_tree(&s, location, (0, 0), window_type)),
-            _ => None,
-        }
-    }
-
-    pub fn with_surfaces<F>(&self, processor: F)
-    where
-        F: FnMut(&WlSurface, &WlSurfaceData) + Copy,
-    {
-        match self {
-            WindowElement::Wayland(w) => w.with_surfaces(processor),
-            #[cfg(feature = "xwayland")]
-            WindowElement::X11(w) => {
-                if let Some(surface) = w.wl_surface() {
-                    with_surfaces_surface_tree(&surface, processor);
-                }
-            }
-            _ => {}
-        }
-    }
-
-    pub fn send_frame<T, F>(
-        &self,
-        output: &Output,
-        time: T,
-        throttle: Option<Duration>,
-        primary_scan_out_output: F,
-    ) where
-        T: Into<Duration>,
-        F: FnMut(&WlSurface, &WlSurfaceData) -> Option<Output> + Copy,
-    {
-        match self {
-            WindowElement::Wayland(w) => w.send_frame(output, time, throttle, primary_scan_out_output),
-            #[cfg(feature = "xwayland")]
-            WindowElement::X11(w) => {
-                if let Some(surface) = w.wl_surface() {
-                    send_frames_surface_tree(&surface, output, time, throttle, primary_scan_out_output);
-                }
-            }
-            _ => {}
-        }
-    }
-
-    pub fn take_presentation_feedback<F1, F2>(
-        &self,
-        output_feedback: &mut OutputPresentationFeedback,
-        primary_scan_out_output: F1,
-        presentation_feedback_flags: F2,
-    ) where
-        F1: FnMut(&WlSurface, &WlSurfaceData) -> Option<Output> + Copy,
-        F2: FnMut(&WlSurface, &WlSurfaceData) -> wp_presentation_feedback::Kind + Copy,
-    {
-        match self {
-            WindowElement::Wayland(w) => w.take_presentation_feedback(
-                output_feedback,
-                primary_scan_out_output,
-                presentation_feedback_flags,
-            ),
-            #[cfg(feature = "xwayland")]
-            WindowElement::X11(w) => {
-                if let Some(surface) = w.wl_surface() {
-                    take_presentation_feedback_surface_tree(
-                        &surface,
-                        output_feedback,
-                        primary_scan_out_output,
-                        presentation_feedback_flags,
-                    );
-                }
-            }
-            _ => {}
-        }
-    }
-
-    #[cfg(feature = "xwayland")]
-    pub fn is_x11(&self) -> bool {
-        matches!(self, WindowElement::X11(_))
-    }
-
-    pub fn is_wayland(&self) -> bool {
-        matches!(self, WindowElement::Wayland(_))
-    }
-
-    pub fn wl_surface(&self) -> Option<WlSurface> {
-        match self {
-            WindowElement::Wayland(w) => w.wl_surface(),
-            #[cfg(feature = "xwayland")]
-            WindowElement::X11(w) => w.wl_surface(),
-            _ => None,
-        }
-    }
-}
-
-impl<R> AsRenderElements<R> for WindowElement
-where
-    R: Renderer + ImportAll,
-    <R as Renderer>::TextureId: 'static,
-{
-    type RenderElement = WaylandSurfaceRenderElement<R>;
-
-    fn render_elements<C: From<Self::RenderElement>>(
-        &self,
-        renderer: &mut R,
-        location: Point<i32, smithay::utils::Physical>,
-        scale: smithay::utils::Scale<f64>,
-    ) -> Vec<C> {
-        match self {
-            WindowElement::Wayland(w) => w.render_elements(renderer, location, scale),
-            #[cfg(feature = "xwayland")]
-            WindowElement::X11(w) => w.render_elements(renderer, location, scale),
-            _ => unreachable!(),
-        }
-    }
-}
-
-struct MoveSurfaceGrab<B: 'static> {
-    start_data: PointerGrabStartData<AnvilState<B>>,
-    window: WindowElement,
-    initial_window_location: Point<i32, Logical>,
-}
-
-impl<BackendData> PointerGrab<AnvilState<BackendData>> for MoveSurfaceGrab<BackendData> {
-    fn motion(
-        &mut self,
-        data: &mut AnvilState<BackendData>,
-        handle: &mut PointerInnerHandle<'_, AnvilState<BackendData>>,
-        _focus: Option<(FocusTarget, Point<i32, Logical>)>,
-        event: &MotionEvent,
-    ) {
-        // While the grab is active, no client has pointer focus
-        handle.motion(data, None, event);
-
-        let delta = event.location - self.start_data.location;
-        let new_location = self.initial_window_location.to_f64() + delta;
-
-        data.space
-            .map_element(self.window.clone(), new_location.to_i32_round(), true);
-    }
-
-    fn button(
-        &mut self,
-        data: &mut AnvilState<BackendData>,
-        handle: &mut PointerInnerHandle<'_, AnvilState<BackendData>>,
-        event: &ButtonEvent,
-    ) {
-        handle.button(data, event);
-        if handle.current_pressed().is_empty() {
-            // No more buttons are pressed, release the grab.
-            handle.unset_grab(data, event.serial, event.time);
-        }
-    }
-
-    fn axis(
-        &mut self,
-        data: &mut AnvilState<BackendData>,
-        handle: &mut PointerInnerHandle<'_, AnvilState<BackendData>>,
-        details: AxisFrame,
-    ) {
-        handle.axis(data, details)
-    }
-
-    fn start_data(&self) -> &PointerGrabStartData<AnvilState<BackendData>> {
-        &self.start_data
-    }
-}
-
-bitflags::bitflags! {
-    struct ResizeEdge: u32 {
-        const NONE = 0;
-        const TOP = 1;
-        const BOTTOM = 2;
-        const LEFT = 4;
-        const TOP_LEFT = 5;
-        const BOTTOM_LEFT = 6;
-        const RIGHT = 8;
-        const TOP_RIGHT = 9;
-        const BOTTOM_RIGHT = 10;
-    }
-}
-
-impl From<xdg_toplevel::ResizeEdge> for ResizeEdge {
-    #[inline]
-    fn from(x: xdg_toplevel::ResizeEdge) -> Self {
-        Self::from_bits(x as u32).unwrap()
-    }
-}
-
-impl From<ResizeEdge> for xdg_toplevel::ResizeEdge {
-    #[inline]
-    fn from(x: ResizeEdge) -> Self {
-        use std::convert::TryFrom;
-
-        Self::try_from(x.bits()).unwrap()
-    }
-}
-
-impl From<X11ResizeEdge> for ResizeEdge {
-    fn from(edge: X11ResizeEdge) -> Self {
-        match edge {
-            X11ResizeEdge::Bottom => ResizeEdge::BOTTOM,
-            X11ResizeEdge::BottomLeft => ResizeEdge::BOTTOM_LEFT,
-            X11ResizeEdge::BottomRight => ResizeEdge::BOTTOM_RIGHT,
-            X11ResizeEdge::Left => ResizeEdge::LEFT,
-            X11ResizeEdge::Right => ResizeEdge::RIGHT,
-            X11ResizeEdge::Top => ResizeEdge::TOP,
-            X11ResizeEdge::TopLeft => ResizeEdge::TOP_LEFT,
-            X11ResizeEdge::TopRight => ResizeEdge::TOP_RIGHT,
-        }
-    }
-}
-
-struct ResizeSurfaceGrab<B: 'static> {
-    start_data: PointerGrabStartData<AnvilState<B>>,
-    window: WindowElement,
-    edges: ResizeEdge,
-    initial_window_location: Point<i32, Logical>,
-    initial_window_size: Size<i32, Logical>,
-    last_window_size: Size<i32, Logical>,
-}
-
-impl<BackendData> PointerGrab<AnvilState<BackendData>> for ResizeSurfaceGrab<BackendData> {
-    fn motion(
-        &mut self,
-        data: &mut AnvilState<BackendData>,
-        handle: &mut PointerInnerHandle<'_, AnvilState<BackendData>>,
-        _focus: Option<(FocusTarget, Point<i32, Logical>)>,
-        event: &MotionEvent,
-    ) {
-        // While the grab is active, no client has pointer focus
-        handle.motion(data, None, event);
-
-        // It is impossible to get `min_size` and `max_size` of dead toplevel, so we return early.
-        if !self.window.alive() {
-            handle.unset_grab(data, event.serial, event.time);
-            return;
-        }
-
-        let (mut dx, mut dy) = (event.location - self.start_data.location).into();
-
-        let mut new_window_width = self.initial_window_size.w;
-        let mut new_window_height = self.initial_window_size.h;
-
-        let left_right = ResizeEdge::LEFT | ResizeEdge::RIGHT;
-        let top_bottom = ResizeEdge::TOP | ResizeEdge::BOTTOM;
-
-        if self.edges.intersects(left_right) {
-            if self.edges.intersects(ResizeEdge::LEFT) {
-                dx = -dx;
-            }
-
-            new_window_width = (self.initial_window_size.w as f64 + dx) as i32;
-        }
-
-        if self.edges.intersects(top_bottom) {
-            if self.edges.intersects(ResizeEdge::TOP) {
-                dy = -dy;
-            }
-
-            new_window_height = (self.initial_window_size.h as f64 + dy) as i32;
-        }
-
-        let (min_size, max_size) = if let Some(surface) = self.window.wl_surface() {
-            with_states(&surface, |states| {
-                let data = states.cached_state.current::<SurfaceCachedState>();
-                (data.min_size, data.max_size)
-            })
-        } else {
-            ((0, 0).into(), (0, 0).into())
-        };
-
-        let min_width = min_size.w.max(1);
-        let min_height = min_size.h.max(1);
-        let max_width = if max_size.w == 0 {
-            i32::max_value()
-        } else {
-            max_size.w
-        };
-        let max_height = if max_size.h == 0 {
-            i32::max_value()
-        } else {
-            max_size.h
-        };
-
-        new_window_width = new_window_width.max(min_width).min(max_width);
-        new_window_height = new_window_height.max(min_height).min(max_height);
-
-        self.last_window_size = (new_window_width, new_window_height).into();
-
-        match &self.window {
-            WindowElement::Wayland(w) => {
-                let xdg = w.toplevel();
-                xdg.with_pending_state(|state| {
-                    state.states.set(xdg_toplevel::State::Resizing);
-                    state.size = Some(self.last_window_size);
-                });
-                xdg.send_configure();
-            }
-            #[cfg(feature = "xwayland")]
-            WindowElement::X11(x11) => {
-                let location = data.space.element_location(&self.window).unwrap();
-                x11.configure(Rectangle::from_loc_and_size(location, self.last_window_size))
-                    .unwrap();
-            }
-            _ => {}
-        }
-    }
-
-    fn button(
-        &mut self,
-        data: &mut AnvilState<BackendData>,
-        handle: &mut PointerInnerHandle<'_, AnvilState<BackendData>>,
-        event: &ButtonEvent,
-    ) {
-        handle.button(data, event);
-        if handle.current_pressed().is_empty() {
-            // No more buttons are pressed, release the grab.
-            handle.unset_grab(data, event.serial, event.time);
-
-            // If toplevel is dead, we can't resize it, so we return early.
-            if !self.window.alive() {
-                return;
-            }
-
-            match &self.window {
-                WindowElement::Wayland(w) => {
-                    let xdg = w.toplevel();
-                    xdg.with_pending_state(|state| {
-                        state.states.unset(xdg_toplevel::State::Resizing);
-                        state.size = Some(self.last_window_size);
-                    });
-                    xdg.send_configure();
-                    if self.edges.intersects(ResizeEdge::TOP_LEFT) {
-                        let geometry = self.window.geometry();
-                        let mut location = data.space.element_location(&self.window).unwrap();
-
-                        if self.edges.intersects(ResizeEdge::LEFT) {
-                            location.x = self.initial_window_location.x
-                                + (self.initial_window_size.w - geometry.size.w);
-                        }
-                        if self.edges.intersects(ResizeEdge::TOP) {
-                            location.y = self.initial_window_location.y
-                                + (self.initial_window_size.h - geometry.size.h);
-                        }
-
-                        data.space.map_element(self.window.clone(), location, true);
-                    }
-
-                    with_states(&self.window.wl_surface().unwrap(), |states| {
-                        let mut data = states
-                            .data_map
-                            .get::<RefCell<SurfaceData>>()
-                            .unwrap()
-                            .borrow_mut();
-                        if let ResizeState::Resizing(resize_data) = data.resize_state {
-                            data.resize_state = ResizeState::WaitingForFinalAck(resize_data, event.serial);
-                        } else {
-                            panic!("invalid resize state: {:?}", data.resize_state);
-                        }
-                    });
-                }
-                WindowElement::X11(x11) => {
-                    let mut location = data.space.element_location(&self.window).unwrap();
-                    if self.edges.intersects(ResizeEdge::TOP_LEFT) {
-                        let geometry = self.window.geometry();
-
-                        if self.edges.intersects(ResizeEdge::LEFT) {
-                            location.x = self.initial_window_location.x
-                                + (self.initial_window_size.w - geometry.size.w);
-                        }
-                        if self.edges.intersects(ResizeEdge::TOP) {
-                            location.y = self.initial_window_location.y
-                                + (self.initial_window_size.h - geometry.size.h);
-                        }
-
-                        data.space.map_element(self.window.clone(), location, true);
-                    }
-                    x11.configure(Rectangle::from_loc_and_size(location, self.last_window_size))
-                        .unwrap();
-
-                    let Some(surface) = self.window.wl_surface() else {
-                        // X11 Window got unmapped, abort
-                        return
-                    };
-                    with_states(&surface, |states| {
-                        let mut data = states
-                            .data_map
-                            .get::<RefCell<SurfaceData>>()
-                            .unwrap()
-                            .borrow_mut();
-                        if let ResizeState::Resizing(resize_data) = data.resize_state {
-                            data.resize_state = ResizeState::WaitingForCommit(resize_data);
-                        } else {
-                            panic!("invalid resize state: {:?}", data.resize_state);
-                        }
-                    });
-                }
-                _ => {}
-            }
-        }
-    }
-
-    fn axis(
-        &mut self,
-        data: &mut AnvilState<BackendData>,
-        handle: &mut PointerInnerHandle<'_, AnvilState<BackendData>>,
-        details: AxisFrame,
-    ) {
-        handle.axis(data, details)
-    }
-
-    fn start_data(&self) -> &PointerGrabStartData<AnvilState<BackendData>> {
-        &self.start_data
-    }
-}
-
-fn fullscreen_output_geometry(
-    wl_surface: &WlSurface,
-    wl_output: Option<&wl_output::WlOutput>,
-    space: &mut Space<WindowElement>,
-) -> Option<Rectangle<i32, Logical>> {
-    // First test if a specific output has been requested
-    // if the requested output is not found ignore the request
-    wl_output
-        .and_then(Output::from_resource)
-        .or_else(|| {
-            let w = space
-                .elements()
-                .find(|window| window.wl_surface().map(|s| s == *wl_surface).unwrap_or(false))
-                .cloned();
-            w.and_then(|w| space.outputs_for_element(&w).get(0).cloned())
-        })
-        .and_then(|o| space.output_geometry(&o))
-}
-
-#[derive(Default)]
-pub struct FullscreenSurface(RefCell<Option<WindowElement>>);
-
-impl FullscreenSurface {
-    pub fn set(&self, window: WindowElement) {
-        *self.0.borrow_mut() = Some(window);
-    }
-
-    pub fn get(&self) -> Option<WindowElement> {
-        self.0.borrow().clone()
-    }
-
-    pub fn clear(&self) -> Option<WindowElement> {
-        self.0.borrow_mut().take()
-    }
-}
-
-impl<BackendData> BufferHandler for AnvilState<BackendData> {
-    fn buffer_destroyed(&mut self, _buffer: &WlBuffer) {}
-}
-
-impl<BackendData: Backend> CompositorHandler for AnvilState<BackendData> {
-    fn compositor_state(&mut self) -> &mut CompositorState {
-        &mut self.compositor_state
-    }
-    fn commit(&mut self, surface: &WlSurface) {
-        #[cfg(feature = "xwayland")]
-        X11WM::commit_hook(surface);
-
-        on_commit_buffer_handler(surface);
-        self.backend_data.early_import(surface);
-
-        if !is_sync_subsurface(surface) {
-            let mut root = surface.clone();
-            while let Some(parent) = get_parent(&root) {
-                root = parent;
-            }
-            if let Some(WindowElement::Wayland(window)) = self.window_for_surface(&root) {
-                window.on_commit();
-            }
-        }
-        self.popups.commit(surface);
-
-        ensure_initial_configure(surface, &self.space, &mut self.popups)
-    }
-}
-
-impl<BackendData: Backend> XdgShellHandler for AnvilState<BackendData> {
-    fn xdg_shell_state(&mut self) -> &mut XdgShellState {
-        &mut self.xdg_shell_state
-    }
-
-    fn new_toplevel(&mut self, surface: ToplevelSurface) {
-        // Do not send a configure here, the initial configure
-        // of a xdg_surface has to be sent during the commit if
-        // the surface is not already configured
-        let window = WindowElement::Wayland(Window::new(surface));
-        place_new_window(&mut self.space, &window, true);
-    }
-
-    fn new_popup(&mut self, surface: PopupSurface, positioner: PositionerState) {
-        // Do not send a configure here, the initial configure
-        // of a xdg_surface has to be sent during the commit if
-        // the surface is not already configured
-
-        // TODO: properly recompute the geometry with the whole of positioner state
-        surface.with_pending_state(|state| {
-            // NOTE: This is not really necessary as the default geometry
-            // is already set the same way, but for demonstrating how
-            // to set the initial popup geometry this code is left as
-            // an example
-            state.geometry = positioner.get_geometry();
-        });
-        if let Err(err) = self.popups.track_popup(PopupKind::from(surface)) {
-            slog::warn!(self.log, "Failed to track popup: {}", err);
-        }
-    }
-
-    fn reposition_request(&mut self, surface: PopupSurface, positioner: PositionerState, token: u32) {
-        surface.with_pending_state(|state| {
-            // NOTE: This is again a simplification, a proper compositor would
-            // calculate the geometry of the popup here. For simplicity we just
-            // use the default implementation here that does not take the
-            // window position and output constraints into account.
-            let geometry = positioner.get_geometry();
-            state.geometry = geometry;
-            state.positioner = positioner;
-        });
-        surface.send_repositioned(token);
-    }
-
-    fn move_request(&mut self, surface: ToplevelSurface, seat: wl_seat::WlSeat, serial: Serial) {
-        let seat: Seat<AnvilState<BackendData>> = Seat::from_resource(&seat).unwrap();
-        // TODO: touch move.
-        let pointer = seat.get_pointer().unwrap();
-
-        // Check that this surface has a click grab.
-        if !pointer.has_grab(serial) {
-            return;
-        }
-
-        let start_data = pointer.grab_start_data().unwrap();
-
-        let window = self.window_for_surface(surface.wl_surface()).unwrap();
-
-        // If the focus was for a different surface, ignore the request.
-        if start_data.focus.is_none()
-            || !start_data
-                .focus
-                .as_ref()
-                .unwrap()
-                .0
-                .same_client_as(&surface.wl_surface().id())
-        {
-            return;
-        }
-
-        let mut initial_window_location = self.space.element_location(&window).unwrap();
-
-        // If surface is maximized then unmaximize it
-        let current_state = surface.current_state();
-        if current_state.states.contains(xdg_toplevel::State::Maximized) {
-            surface.with_pending_state(|state| {
-                state.states.unset(xdg_toplevel::State::Maximized);
-                state.size = None;
-            });
-
-            surface.send_configure();
-
-            // NOTE: In real compositor mouse location should be mapped to a new window size
-            // For example, you could:
-            // 1) transform mouse pointer position from compositor space to window space (location relative)
-            // 2) divide the x coordinate by width of the window to get the percentage
-            //   - 0.0 would be on the far left of the window
-            //   - 0.5 would be in middle of the window
-            //   - 1.0 would be on the far right of the window
-            // 3) multiply the percentage by new window width
-            // 4) by doing that, drag will look a lot more natural
-            //
-            // but for anvil needs setting location to pointer location is fine
-            let pos = pointer.current_location();
-            initial_window_location = (pos.x as i32, pos.y as i32).into();
-        }
-
-        let grab = MoveSurfaceGrab {
-            start_data,
-            window,
-            initial_window_location,
-        };
-
-        pointer.set_grab(self, grab, serial, Focus::Clear);
-    }
-
-    fn resize_request(
-        &mut self,
-        surface: ToplevelSurface,
-        seat: wl_seat::WlSeat,
-        serial: Serial,
-        edges: xdg_toplevel::ResizeEdge,
-    ) {
-        let seat: Seat<AnvilState<BackendData>> = Seat::from_resource(&seat).unwrap();
-        // TODO: touch resize.
-        let pointer = seat.get_pointer().unwrap();
-
-        // Check that this surface has a click grab.
-        if !pointer.has_grab(serial) {
-            return;
-        }
-
-        let start_data = pointer.grab_start_data().unwrap();
-
-        let window = self.window_for_surface(surface.wl_surface()).unwrap();
-
-        // If the focus was for a different surface, ignore the request.
-        if start_data.focus.is_none()
-            || !start_data
-                .focus
-                .as_ref()
-                .unwrap()
-                .0
-                .same_client_as(&surface.wl_surface().id())
-        {
-            return;
-        }
-
-        let geometry = window.geometry();
-        let loc = self.space.element_location(&window).unwrap();
-        let (initial_window_location, initial_window_size) = (loc, geometry.size);
-
-        with_states(surface.wl_surface(), move |states| {
-            states
-                .data_map
-                .get::<RefCell<SurfaceData>>()
-                .unwrap()
-                .borrow_mut()
-                .resize_state = ResizeState::Resizing(ResizeData {
-                edges: edges.into(),
-                initial_window_location,
-                initial_window_size,
-            });
-        });
-
-        let grab = ResizeSurfaceGrab {
-            start_data,
-            window,
-            edges: edges.into(),
-            initial_window_location,
-            initial_window_size,
-            last_window_size: initial_window_size,
-        };
-
-        pointer.set_grab(self, grab, serial, Focus::Clear);
-    }
-
-    fn ack_configure(&mut self, surface: WlSurface, configure: Configure) {
-        if let Configure::Toplevel(configure) = configure {
-            if let Some(serial) = with_states(&surface, |states| {
-                if let Some(data) = states.data_map.get::<RefCell<SurfaceData>>() {
-                    if let ResizeState::WaitingForFinalAck(_, serial) = data.borrow().resize_state {
-                        return Some(serial);
-                    }
-                }
-
-                None
-            }) {
-                // When the resize grab is released the surface
-                // resize state will be set to WaitingForFinalAck
-                // and the client will receive a configure request
-                // without the resize state to inform the client
-                // resizing has finished. Here we will wait for
-                // the client to acknowledge the end of the
-                // resizing. To check if the surface was resizing
-                // before sending the configure we need to use
-                // the current state as the received acknowledge
-                // will no longer have the resize state set
-                let is_resizing = with_states(&surface, |states| {
-                    states
-                        .data_map
-                        .get::<XdgToplevelSurfaceData>()
-                        .unwrap()
-                        .lock()
-                        .unwrap()
-                        .current
-                        .states
-                        .contains(xdg_toplevel::State::Resizing)
-                });
-
-                if configure.serial >= serial && is_resizing {
-                    with_states(&surface, |states| {
-                        let mut data = states
-                            .data_map
-                            .get::<RefCell<SurfaceData>>()
-                            .unwrap()
-                            .borrow_mut();
-                        if let ResizeState::WaitingForFinalAck(resize_data, _) = data.resize_state {
-                            data.resize_state = ResizeState::WaitingForCommit(resize_data);
-                        } else {
-                            unreachable!()
-                        }
-                    });
-                }
-            }
-        }
-    }
-
-    fn fullscreen_request(&mut self, surface: ToplevelSurface, mut wl_output: Option<wl_output::WlOutput>) {
-        // NOTE: This is only one part of the solution. We can set the
-        // location and configure size here, but the surface should be rendered fullscreen
-        // independently from its buffer size
-        let wl_surface = surface.wl_surface();
-
-        let output_geometry = fullscreen_output_geometry(wl_surface, wl_output.as_ref(), &mut self.space);
-
-        if let Some(geometry) = output_geometry {
-            let output = wl_output
-                .as_ref()
-                .and_then(Output::from_resource)
-                .unwrap_or_else(|| self.space.outputs().next().unwrap().clone());
-            let client = self.display_handle.get_client(wl_surface.id()).unwrap();
-            for output in output.client_outputs(&client) {
-                wl_output = Some(output);
-            }
-            let window = self
-                .space
-                .elements()
-                .find(|window| window.wl_surface().map(|s| s == *wl_surface).unwrap_or(false))
-                .unwrap();
-
-            surface.with_pending_state(|state| {
-                state.states.set(xdg_toplevel::State::Fullscreen);
-                state.size = Some(geometry.size);
-                state.fullscreen_output = wl_output;
-            });
-            surface.send_configure();
-            output.user_data().insert_if_missing(FullscreenSurface::default);
-            output
-                .user_data()
-                .get::<FullscreenSurface>()
-                .unwrap()
-                .set(window.clone());
-            slog::trace!(self.log, "Fullscreening: {:?}", window);
-        }
-    }
-
-    fn unfullscreen_request(&mut self, surface: ToplevelSurface) {
-        let ret = surface.with_pending_state(|state| {
-            state.states.unset(xdg_toplevel::State::Fullscreen);
-            state.size = None;
-            state.fullscreen_output.take()
-        });
-        if let Some(output) = ret {
-            let output = Output::from_resource(&output).unwrap();
-            if let Some(fullscreen) = output.user_data().get::<FullscreenSurface>() {
-                slog::trace!(self.log, "Unfullscreening: {:?}", fullscreen.get());
-                fullscreen.clear();
-                self.backend_data.reset_buffers(&output);
-            }
-
-            surface.send_configure();
-        }
-    }
-
-    fn maximize_request(&mut self, surface: ToplevelSurface) {
-        // NOTE: This should use layer-shell when it is implemented to
-        // get the correct maximum size
-        let window = self.window_for_surface(surface.wl_surface()).unwrap();
-        let outputs_for_window = self.space.outputs_for_element(&window);
-        let output = outputs_for_window
-            .first()
-            // The window hasn't been mapped yet, use the primary output instead
-            .or_else(|| self.space.outputs().next())
-            // Assumes that at least one output exists
-            .expect("No outputs found");
-        let geometry = self.space.output_geometry(output).unwrap();
-
-        surface.with_pending_state(|state| {
-            state.states.set(xdg_toplevel::State::Maximized);
-            state.size = Some(geometry.size);
-        });
-        surface.send_configure();
-        self.space.map_element(window, geometry.loc, true);
-    }
-
-    fn unmaximize_request(&mut self, surface: ToplevelSurface) {
-        surface.with_pending_state(|state| {
-            state.states.unset(xdg_toplevel::State::Maximized);
-            state.size = None;
-        });
-        surface.send_configure();
-    }
-
-    fn grab(&mut self, surface: PopupSurface, seat: wl_seat::WlSeat, serial: Serial) {
-        let seat: Seat<AnvilState<BackendData>> = Seat::from_resource(&seat).unwrap();
-        let kind = PopupKind::Xdg(surface);
-        if let Some(root) = find_popup_root_surface(&kind).ok().and_then(|root| {
-            self.space
-                .elements()
-                .find(|w| w.wl_surface().map(|s| s == root).unwrap_or(false))
-                .cloned()
-                .map(FocusTarget::Window)
-                .or_else(|| {
-                    self.space
-                        .outputs()
-                        .find_map(|o| {
-                            let map = layer_map_for_output(o);
-                            map.layer_for_surface(&root, WindowSurfaceType::TOPLEVEL).cloned()
-                        })
-                        .map(FocusTarget::LayerSurface)
-                })
-        }) {
-            let ret = self.popups.grab_popup(root, kind, &seat, serial);
-
-            if let Ok(mut grab) = ret {
-                if let Some(keyboard) = seat.get_keyboard() {
-                    if keyboard.is_grabbed()
-                        && !(keyboard.has_grab(serial)
-                            || keyboard.has_grab(grab.previous_serial().unwrap_or(serial)))
-                    {
-                        grab.ungrab(PopupUngrabStrategy::All);
-                        return;
-                    }
-                    keyboard.set_focus(self, grab.current_grab(), serial);
-                    keyboard.set_grab(PopupKeyboardGrab::new(&grab), serial);
-                }
-                if let Some(pointer) = seat.get_pointer() {
-                    if pointer.is_grabbed()
-                        && !(pointer.has_grab(serial)
-                            || pointer.has_grab(grab.previous_serial().unwrap_or_else(|| grab.serial())))
-                    {
-                        grab.ungrab(PopupUngrabStrategy::All);
-                        return;
-                    }
-                    pointer.set_grab(self, PopupPointerGrab::new(&grab), serial, Focus::Keep);
-                }
-            }
-        }
-    }
-}
-
-impl<BackendData> WlrLayerShellHandler for AnvilState<BackendData> {
-    fn shell_state(&mut self) -> &mut WlrLayerShellState {
-        &mut self.layer_shell_state
-    }
-
-    fn new_layer_surface(
-        &mut self,
-        surface: WlrLayerSurface,
-        wl_output: Option<wl_output::WlOutput>,
-        _layer: Layer,
-        namespace: String,
-    ) {
-        let output = wl_output
-            .as_ref()
-            .and_then(Output::from_resource)
-            .unwrap_or_else(|| self.space.outputs().next().unwrap().clone());
-        let mut map = layer_map_for_output(&output);
-        map.map_layer(&LayerSurface::new(surface, namespace)).unwrap();
-    }
-}
-
-#[derive(Debug)]
-struct OldGeometry(RefCell<Option<Rectangle<i32, Logical>>>);
-impl Default for OldGeometry {
-    fn default() -> Self {
-        OldGeometry(RefCell::new(None))
-    }
-}
-impl OldGeometry {
-    pub fn save(&self, geo: Rectangle<i32, Logical>) {
-        *self.0.borrow_mut() = Some(geo);
-    }
-
-    pub fn restore(&self) -> Option<Rectangle<i32, Logical>> {
-        self.0.borrow_mut().take()
-    }
-}
-
-impl<BackendData: Backend> XwmHandler for CalloopData<BackendData> {
-    fn xwm_state(&mut self, _xwm: XwmId) -> &mut X11WM {
-        self.state.xwm.as_mut().unwrap()
-    }
-
-    fn new_window(&mut self, _xwm: XwmId, _window: X11Surface) {}
-    fn new_override_redirect_window(&mut self, _xwm: XwmId, _window: X11Surface) {}
-    fn map_window_request(&mut self, _xwm: XwmId, window: X11Surface) {
-        window.set_mapped(true).unwrap();
-        let window = WindowElement::X11(window);
-        place_new_window(&mut self.state.space, &window, true);
-        let bbox = self.state.space.element_bbox(&window).unwrap();
-        let WindowElement::X11(xsurface) = window else { unreachable!() };
-        xsurface.configure(Some(bbox)).unwrap();
-    }
-    fn mapped_override_redirect_window(&mut self, _xwm: XwmId, window: X11Surface) {
-        let location = window.geometry().loc;
-        let window = WindowElement::X11(window);
-        self.state.space.map_element(window, location, true);
-    }
-    fn unmapped_window(&mut self, _xwm: XwmId, window: X11Surface) {
-        let maybe = self
-            .state
-            .space
-            .elements()
-            .find(|e| matches!(e, WindowElement::X11(w) if w == &window))
-            .cloned();
-        if let Some(elem) = maybe {
-            self.state.space.unmap_elem(&elem)
-        }
-        if !window.is_override_redirect() {
-            window.set_mapped(false).unwrap();
-        }
-    }
-    fn destroyed_window(&mut self, _xwm: XwmId, _window: X11Surface) {}
-
-    fn configure_request(
-        &mut self,
-        _xwm: XwmId,
-        _window: X11Surface,
-        _x: Option<i32>,
-        _y: Option<i32>,
-        _w: Option<u32>,
-        _h: Option<u32>,
-    ) {
-        // Nope
-    }
-    fn configure_notify(&mut self, _xwm: XwmId, window: X11Surface, x: i32, y: i32, _w: u32, _h: u32) {
-        let Some(elem) = self
-            .state
-            .space
-            .elements()
-            .find(|e| matches!(e, WindowElement::X11(w) if w == &window))
-            .cloned()
-        else { return };
-        self.state.space.map_element(elem, (x, y), false);
-    }
-
-    fn maximize_request(&mut self, _xwm: XwmId, window: X11Surface) {
-        let Some(elem) = self
-            .state
-            .space
-            .elements()
-            .find(|e| matches!(e, WindowElement::X11(w) if w == &window))
-            .cloned()
-        else { return };
-
-        let old_geo = self.state.space.element_bbox(&elem).unwrap();
-        let outputs_for_window = self.state.space.outputs_for_element(&elem);
-        let output = outputs_for_window
-            .first()
-            // The window hasn't been mapped yet, use the primary output instead
-            .or_else(|| self.state.space.outputs().next())
-            // Assumes that at least one output exists
-            .expect("No outputs found");
-        let geometry = self.state.space.output_geometry(output).unwrap();
-
-        window.set_maximized(true).unwrap();
-        window.configure(geometry).unwrap();
-        window.user_data().insert_if_missing(OldGeometry::default);
-        window.user_data().get::<OldGeometry>().unwrap().save(old_geo);
-        self.state.space.map_element(elem, geometry.loc, false);
-    }
-    fn unmaximize_request(&mut self, _xwm: XwmId, window: X11Surface) {
-        let Some(elem) = self
-            .state
-            .space
-            .elements()
-            .find(|e| matches!(e, WindowElement::X11(w) if w == &window))
-            .cloned()
-        else { return };
-
-        window.set_maximized(false).unwrap();
-        if let Some(old_geo) = window
-            .user_data()
-            .get::<OldGeometry>()
-            .and_then(|data| data.restore())
-        {
-            window.configure(old_geo).unwrap();
-            self.state.space.map_element(elem, old_geo.loc, false);
-        }
-    }
-    fn fullscreen_request(&mut self, _xwm: XwmId, window: X11Surface) {
-        if let Some(elem) = self
-            .state
-            .space
-            .elements()
-            .find(|e| matches!(e, WindowElement::X11(w) if w == &window))
-        {
-            let outputs_for_window = self.state.space.outputs_for_element(elem);
-            let output = outputs_for_window
-                .first()
-                // The window hasn't been mapped yet, use the primary output instead
-                .or_else(|| self.state.space.outputs().next())
-                // Assumes that at least one output exists
-                .expect("No outputs found");
-            let geometry = self.state.space.output_geometry(output).unwrap();
-
-            window.set_fullscreen(true).unwrap();
-            window.configure(geometry).unwrap();
-            output.user_data().insert_if_missing(FullscreenSurface::default);
-            output
-                .user_data()
-                .get::<FullscreenSurface>()
-                .unwrap()
-                .set(elem.clone());
-            slog::trace!(self.state.log, "Fullscreening: {:?}", elem);
-        }
-    }
-    fn unfullscreen_request(&mut self, _xwm: XwmId, window: X11Surface) {
-        if let Some(elem) = self
-            .state
-            .space
-            .elements()
-            .find(|e| matches!(e, WindowElement::X11(w) if w == &window))
-        {
-            window.set_fullscreen(false).unwrap();
-            if let Some(output) = self.state.space.outputs().find(|o| {
-                o.user_data()
-                    .get::<FullscreenSurface>()
-                    .and_then(|f| f.get())
-                    .map(|w| &w == elem)
-                    .unwrap_or(false)
-            }) {
-                slog::trace!(self.state.log, "Unfullscreening: {:?}", elem);
-                output.user_data().get::<FullscreenSurface>().unwrap().clear();
-                window.configure(self.state.space.element_bbox(elem)).unwrap();
-                self.state.backend_data.reset_buffers(output);
-            }
-        }
-    }
-
-    fn resize_request(&mut self, _xwm: XwmId, window: X11Surface, _button: u32, edges: X11ResizeEdge) {
-        let seat = &self.state.seat; // luckily anvil only supports one seat anyway...
-        let pointer = seat.get_pointer().unwrap();
-        let start_data = pointer.grab_start_data().unwrap();
-
-        let Some(element) = self
-            .state
-            .space
-            .elements()
-            .find(|e| matches!(e, WindowElement::X11(w) if w == &window)) else { return };
-
-        let geometry = element.geometry();
-        let loc = self.state.space.element_location(element).unwrap();
-        let (initial_window_location, initial_window_size) = (loc, geometry.size);
-
-        with_states(&element.wl_surface().unwrap(), move |states| {
-            states
-                .data_map
-                .get::<RefCell<SurfaceData>>()
-                .unwrap()
-                .borrow_mut()
-                .resize_state = ResizeState::Resizing(ResizeData {
-                edges: edges.into(),
-                initial_window_location,
-                initial_window_size,
-            });
-        });
-
-        let grab = ResizeSurfaceGrab {
-            start_data,
-            window: element.clone(),
-            edges: edges.into(),
-            initial_window_location,
-            initial_window_size,
-            last_window_size: initial_window_size,
-        };
-
-        pointer.set_grab(&mut self.state, grab, SERIAL_COUNTER.next_serial(), Focus::Clear);
-    }
-
-    fn move_request(&mut self, _xwm: XwmId, window: X11Surface, _button: u32) {
-        let seat = &self.state.seat; // luckily anvil only supports one seat anyway...
-        let pointer = seat.get_pointer().unwrap();
-        let start_data = pointer.grab_start_data().unwrap();
-
-        let Some(element) = self
-            .state
-            .space
-            .elements()
-            .find(|e| matches!(e, WindowElement::X11(w) if w == &window)) else { return };
-
-        let mut initial_window_location = self.state.space.element_location(element).unwrap();
-
-        // If surface is maximized then unmaximize it
-        if window.is_maximized() {
-            window.set_maximized(false).unwrap();
-            let pos = pointer.current_location();
-            initial_window_location = (pos.x as i32, pos.y as i32).into();
-            if let Some(old_geo) = window
-                .user_data()
-                .get::<OldGeometry>()
-                .and_then(|data| data.restore())
-            {
-                window
-                    .configure(Rectangle::from_loc_and_size(
-                        initial_window_location,
-                        old_geo.size,
-                    ))
-                    .unwrap();
-            }
-        }
-
-        let grab = MoveSurfaceGrab {
-            start_data,
-            window: element.clone(),
-            initial_window_location,
-        };
-
-        pointer.set_grab(&mut self.state, grab, SERIAL_COUNTER.next_serial(), Focus::Clear);
-    }
-}
-
-impl<BackendData> AnvilState<BackendData> {
-    pub fn window_for_surface(&self, surface: &WlSurface) -> Option<WindowElement> {
-        self.space
-            .elements()
-            .find(|window| window.wl_surface().map(|s| s == *surface).unwrap_or(false))
-            .cloned()
-    }
-}
-
-/// Information about the resize operation.
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-pub struct ResizeData {
-    /// The edges the surface is being resized with.
-    edges: ResizeEdge,
-    /// The initial window location.
-    initial_window_location: Point<i32, Logical>,
-    /// The initial window size (geometry width and height).
-    initial_window_size: Size<i32, Logical>,
-}
-
-/// State of the resize operation.
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-pub enum ResizeState {
-    /// The surface is not being resized.
-    NotResizing,
-    /// The surface is currently being resized.
-    Resizing(ResizeData),
-    /// The resize has finished, and the surface needs to ack the final configure.
-    WaitingForFinalAck(ResizeData, Serial),
-    /// The resize has finished, and the surface needs to commit its final state.
-    WaitingForCommit(ResizeData),
-}
-
-impl Default for ResizeState {
-    fn default() -> Self {
-        ResizeState::NotResizing
-    }
-}
-
-#[derive(Default)]
-pub struct SurfaceData {
-    pub geometry: Option<Rectangle<i32, Logical>>,
-    pub resize_state: ResizeState,
-}
-
-fn ensure_initial_configure(surface: &WlSurface, space: &Space<WindowElement>, popups: &mut PopupManager) {
-    with_surface_tree_upward(
-        surface,
-        (),
-        |_, _, _| TraversalAction::DoChildren(()),
-        |_, states, _| {
-            states
-                .data_map
-                .insert_if_missing(|| RefCell::new(SurfaceData::default()));
-        },
-        |_, _, _| true,
-    );
-
-    if let Some(window) = space
-        .elements()
-        .find(|window| window.wl_surface().map(|s| s == *surface).unwrap_or(false))
-        .cloned()
-    {
-        // send the initial configure if relevant
-        #[cfg_attr(not(feature = "xwayland"), allow(irrefutable_let_patterns))]
-        if let WindowElement::Wayland(ref toplevel) = window {
-            let initial_configure_sent = with_states(surface, |states| {
-                states
-                    .data_map
-                    .get::<XdgToplevelSurfaceData>()
-                    .unwrap()
-                    .lock()
-                    .unwrap()
-                    .initial_configure_sent
-            });
-            if !initial_configure_sent {
-                toplevel.toplevel().send_configure();
-            }
-        }
-
-        with_states(surface, |states| {
-            let mut data = states
-                .data_map
-                .get::<RefCell<SurfaceData>>()
-                .unwrap()
-                .borrow_mut();
-
-            // Finish resizing.
-            if let ResizeState::WaitingForCommit(_) = data.resize_state {
-                data.resize_state = ResizeState::NotResizing;
-            }
-        });
-
-        return;
-    }
-
-    if let Some(popup) = popups.find_popup(surface) {
-        let PopupKind::Xdg(ref popup) = popup;
-        let initial_configure_sent = with_states(surface, |states| {
-            states
-                .data_map
-                .get::<XdgPopupSurfaceData>()
-                .unwrap()
-                .lock()
-                .unwrap()
-                .initial_configure_sent
-        });
-        if !initial_configure_sent {
-            // NOTE: This should never fail as the initial configure is always
-            // allowed.
-            popup.send_configure().expect("initial configure failed");
-        }
-
-        return;
-    };
-
-    if let Some(output) = space.outputs().find(|o| {
-        let map = layer_map_for_output(o);
-        map.layer_for_surface(surface, WindowSurfaceType::TOPLEVEL)
-            .is_some()
-    }) {
-        let initial_configure_sent = with_states(surface, |states| {
-            states
-                .data_map
-                .get::<LayerSurfaceData>()
-                .unwrap()
-                .lock()
-                .unwrap()
-                .initial_configure_sent
-        });
-
-        // send the initial configure if relevant
-        if !initial_configure_sent {
-            let mut map = layer_map_for_output(output);
-
-            // arrange the layers before sending the initial configure
-            // to respect any size the client may have sent
-            map.arrange();
-
-            let layer = map
-                .layer_for_surface(surface, WindowSurfaceType::TOPLEVEL)
-                .unwrap();
-
-            layer.layer_surface().send_configure();
-        }
-    };
-}
-
-fn place_new_window(space: &mut Space<WindowElement>, window: &WindowElement, activate: bool) {
-    // place the window at a random location on the primary output
-    // or if there is not output in a [0;800]x[0;800] square
-    use rand::distributions::{Distribution, Uniform};
-
-    let output = space.outputs().next().cloned();
-    let output_geometry = output
-        .and_then(|o| {
-            let geo = space.output_geometry(&o)?;
-            let map = layer_map_for_output(&o);
-            let zone = map.non_exclusive_zone();
-            Some(Rectangle::from_loc_and_size(geo.loc + zone.loc, zone.size))
-        })
-        .unwrap_or_else(|| Rectangle::from_loc_and_size((0, 0), (800, 800)));
-
-    let max_x = output_geometry.loc.x + (((output_geometry.size.w as f32) / 3.0) * 2.0) as i32;
-    let max_y = output_geometry.loc.y + (((output_geometry.size.h as f32) / 3.0) * 2.0) as i32;
-    let x_range = Uniform::new(output_geometry.loc.x, max_x);
-    let y_range = Uniform::new(output_geometry.loc.y, max_y);
-    let mut rng = rand::thread_rng();
-    let x = x_range.sample(&mut rng);
-    let y = y_range.sample(&mut rng);
-
-    space.map_element(window.clone(), (x, y), activate);
-}
-
-pub fn fixup_positions(space: &mut Space<WindowElement>) {
-    // fixup outputs
-    let mut offset = Point::<i32, Logical>::from((0, 0));
-    for output in space.outputs().cloned().collect::<Vec<_>>().into_iter() {
-        let size = space
-            .output_geometry(&output)
-            .map(|geo| geo.size)
-            .unwrap_or_else(|| Size::from((0, 0)));
-        space.map_output(&output, offset);
-        layer_map_for_output(&output).arrange();
-        offset.x += size.w;
-    }
-
-    // fixup windows
-    let mut orphaned_windows = Vec::new();
-    let outputs = space
-        .outputs()
-        .flat_map(|o| {
-            let geo = space.output_geometry(o)?;
-            let map = layer_map_for_output(o);
-            let zone = map.non_exclusive_zone();
-            Some(Rectangle::from_loc_and_size(geo.loc + zone.loc, zone.size))
-        })
-        .collect::<Vec<_>>();
-    for window in space.elements() {
-        let window_location = match space.element_location(window) {
-            Some(loc) => loc,
-            None => continue,
-        };
-        let geo_loc = window.bbox().loc + window_location;
-
-        if !outputs.iter().any(|o_geo| o_geo.contains(geo_loc)) {
-            orphaned_windows.push(window.clone());
-        }
-    }
-    for window in orphaned_windows.into_iter() {
-        place_new_window(space, &window, false);
-    }
-}
