+++ anvil/src/shell/mod.rs
@@ -1,0 +1,312 @@
+use std::cell::RefCell;
+
+use smithay::{
+    backend::renderer::utils::on_commit_buffer_handler,
+    desktop::{
+        layer_map_for_output, space::SpaceElement, LayerSurface, PopupKind, PopupManager, Space,
+        WindowSurfaceType,
+    },
+    output::Output,
+    reexports::wayland_server::protocol::{wl_buffer::WlBuffer, wl_output, wl_surface::WlSurface},
+    utils::{Logical, Point, Rectangle, Size},
+    wayland::{
+        buffer::BufferHandler,
+        compositor::{
+            get_parent, is_sync_subsurface, with_states, with_surface_tree_upward, CompositorHandler,
+            CompositorState, TraversalAction,
+        },
+        shell::{
+            wlr_layer::{
+                Layer, LayerSurface as WlrLayerSurface, LayerSurfaceData, WlrLayerShellHandler,
+                WlrLayerShellState,
+            },
+            xdg::{XdgPopupSurfaceData, XdgToplevelSurfaceData},
+        },
+    },
+    xwayland::X11WM,
+};
+
+use crate::state::{AnvilState, Backend};
+
+mod element;
+mod grabs;
+#[cfg(feature = "xwayland")]
+mod x11;
+mod xdg;
+
+pub use self::element::*;
+pub use self::grabs::*;
+#[cfg(feature = "xwayland")]
+pub use self::x11::*;
+pub use self::xdg::*;
+
+fn fullscreen_output_geometry(
+    wl_surface: &WlSurface,
+    wl_output: Option<&wl_output::WlOutput>,
+    space: &mut Space<WindowElement>,
+) -> Option<Rectangle<i32, Logical>> {
+    // First test if a specific output has been requested
+    // if the requested output is not found ignore the request
+    wl_output
+        .and_then(Output::from_resource)
+        .or_else(|| {
+            let w = space
+                .elements()
+                .find(|window| window.wl_surface().map(|s| s == *wl_surface).unwrap_or(false))
+                .cloned();
+            w.and_then(|w| space.outputs_for_element(&w).get(0).cloned())
+        })
+        .and_then(|o| space.output_geometry(&o))
+}
+
+#[derive(Default)]
+pub struct FullscreenSurface(RefCell<Option<WindowElement>>);
+
+impl FullscreenSurface {
+    pub fn set(&self, window: WindowElement) {
+        *self.0.borrow_mut() = Some(window);
+    }
+
+    pub fn get(&self) -> Option<WindowElement> {
+        self.0.borrow().clone()
+    }
+
+    pub fn clear(&self) -> Option<WindowElement> {
+        self.0.borrow_mut().take()
+    }
+}
+
+impl<BackendData> BufferHandler for AnvilState<BackendData> {
+    fn buffer_destroyed(&mut self, _buffer: &WlBuffer) {}
+}
+
+impl<BackendData: Backend> CompositorHandler for AnvilState<BackendData> {
+    fn compositor_state(&mut self) -> &mut CompositorState {
+        &mut self.compositor_state
+    }
+    fn commit(&mut self, surface: &WlSurface) {
+        #[cfg(feature = "xwayland")]
+        X11WM::commit_hook(surface);
+
+        on_commit_buffer_handler(surface);
+        self.backend_data.early_import(surface);
+
+        if !is_sync_subsurface(surface) {
+            let mut root = surface.clone();
+            while let Some(parent) = get_parent(&root) {
+                root = parent;
+            }
+            if let Some(WindowElement::Wayland(window)) = self.window_for_surface(&root) {
+                window.on_commit();
+            }
+        }
+        self.popups.commit(surface);
+
+        ensure_initial_configure(surface, &self.space, &mut self.popups)
+    }
+}
+
+impl<BackendData> WlrLayerShellHandler for AnvilState<BackendData> {
+    fn shell_state(&mut self) -> &mut WlrLayerShellState {
+        &mut self.layer_shell_state
+    }
+
+    fn new_layer_surface(
+        &mut self,
+        surface: WlrLayerSurface,
+        wl_output: Option<wl_output::WlOutput>,
+        _layer: Layer,
+        namespace: String,
+    ) {
+        let output = wl_output
+            .as_ref()
+            .and_then(Output::from_resource)
+            .unwrap_or_else(|| self.space.outputs().next().unwrap().clone());
+        let mut map = layer_map_for_output(&output);
+        map.map_layer(&LayerSurface::new(surface, namespace)).unwrap();
+    }
+}
+
+impl<BackendData> AnvilState<BackendData> {
+    pub fn window_for_surface(&self, surface: &WlSurface) -> Option<WindowElement> {
+        self.space
+            .elements()
+            .find(|window| window.wl_surface().map(|s| s == *surface).unwrap_or(false))
+            .cloned()
+    }
+}
+
+#[derive(Default)]
+pub struct SurfaceData {
+    pub geometry: Option<Rectangle<i32, Logical>>,
+    pub resize_state: ResizeState,
+}
+
+fn ensure_initial_configure(surface: &WlSurface, space: &Space<WindowElement>, popups: &mut PopupManager) {
+    with_surface_tree_upward(
+        surface,
+        (),
+        |_, _, _| TraversalAction::DoChildren(()),
+        |_, states, _| {
+            states
+                .data_map
+                .insert_if_missing(|| RefCell::new(SurfaceData::default()));
+        },
+        |_, _, _| true,
+    );
+
+    if let Some(window) = space
+        .elements()
+        .find(|window| window.wl_surface().map(|s| s == *surface).unwrap_or(false))
+        .cloned()
+    {
+        // send the initial configure if relevant
+        #[cfg_attr(not(feature = "xwayland"), allow(irrefutable_let_patterns))]
+        if let WindowElement::Wayland(ref toplevel) = window {
+            let initial_configure_sent = with_states(surface, |states| {
+                states
+                    .data_map
+                    .get::<XdgToplevelSurfaceData>()
+                    .unwrap()
+                    .lock()
+                    .unwrap()
+                    .initial_configure_sent
+            });
+            if !initial_configure_sent {
+                toplevel.toplevel().send_configure();
+            }
+        }
+
+        with_states(surface, |states| {
+            let mut data = states
+                .data_map
+                .get::<RefCell<SurfaceData>>()
+                .unwrap()
+                .borrow_mut();
+
+            // Finish resizing.
+            if let ResizeState::WaitingForCommit(_) = data.resize_state {
+                data.resize_state = ResizeState::NotResizing;
+            }
+        });
+
+        return;
+    }
+
+    if let Some(popup) = popups.find_popup(surface) {
+        let PopupKind::Xdg(ref popup) = popup;
+        let initial_configure_sent = with_states(surface, |states| {
+            states
+                .data_map
+                .get::<XdgPopupSurfaceData>()
+                .unwrap()
+                .lock()
+                .unwrap()
+                .initial_configure_sent
+        });
+        if !initial_configure_sent {
+            // NOTE: This should never fail as the initial configure is always
+            // allowed.
+            popup.send_configure().expect("initial configure failed");
+        }
+
+        return;
+    };
+
+    if let Some(output) = space.outputs().find(|o| {
+        let map = layer_map_for_output(o);
+        map.layer_for_surface(surface, WindowSurfaceType::TOPLEVEL)
+            .is_some()
+    }) {
+        let initial_configure_sent = with_states(surface, |states| {
+            states
+                .data_map
+                .get::<LayerSurfaceData>()
+                .unwrap()
+                .lock()
+                .unwrap()
+                .initial_configure_sent
+        });
+
+        // send the initial configure if relevant
+        if !initial_configure_sent {
+            let mut map = layer_map_for_output(output);
+
+            // arrange the layers before sending the initial configure
+            // to respect any size the client may have sent
+            map.arrange();
+
+            let layer = map
+                .layer_for_surface(surface, WindowSurfaceType::TOPLEVEL)
+                .unwrap();
+
+            layer.layer_surface().send_configure();
+        }
+    };
+}
+
+fn place_new_window(space: &mut Space<WindowElement>, window: &WindowElement, activate: bool) {
+    // place the window at a random location on the primary output
+    // or if there is not output in a [0;800]x[0;800] square
+    use rand::distributions::{Distribution, Uniform};
+
+    let output = space.outputs().next().cloned();
+    let output_geometry = output
+        .and_then(|o| {
+            let geo = space.output_geometry(&o)?;
+            let map = layer_map_for_output(&o);
+            let zone = map.non_exclusive_zone();
+            Some(Rectangle::from_loc_and_size(geo.loc + zone.loc, zone.size))
+        })
+        .unwrap_or_else(|| Rectangle::from_loc_and_size((0, 0), (800, 800)));
+
+    let max_x = output_geometry.loc.x + (((output_geometry.size.w as f32) / 3.0) * 2.0) as i32;
+    let max_y = output_geometry.loc.y + (((output_geometry.size.h as f32) / 3.0) * 2.0) as i32;
+    let x_range = Uniform::new(output_geometry.loc.x, max_x);
+    let y_range = Uniform::new(output_geometry.loc.y, max_y);
+    let mut rng = rand::thread_rng();
+    let x = x_range.sample(&mut rng);
+    let y = y_range.sample(&mut rng);
+
+    space.map_element(window.clone(), (x, y), activate);
+}
+
+pub fn fixup_positions(space: &mut Space<WindowElement>) {
+    // fixup outputs
+    let mut offset = Point::<i32, Logical>::from((0, 0));
+    for output in space.outputs().cloned().collect::<Vec<_>>().into_iter() {
+        let size = space
+            .output_geometry(&output)
+            .map(|geo| geo.size)
+            .unwrap_or_else(|| Size::from((0, 0)));
+        space.map_output(&output, offset);
+        layer_map_for_output(&output).arrange();
+        offset.x += size.w;
+    }
+
+    // fixup windows
+    let mut orphaned_windows = Vec::new();
+    let outputs = space
+        .outputs()
+        .flat_map(|o| {
+            let geo = space.output_geometry(o)?;
+            let map = layer_map_for_output(o);
+            let zone = map.non_exclusive_zone();
+            Some(Rectangle::from_loc_and_size(geo.loc + zone.loc, zone.size))
+        })
+        .collect::<Vec<_>>();
+    for window in space.elements() {
+        let window_location = match space.element_location(window) {
+            Some(loc) => loc,
+            None => continue,
+        };
+        let geo_loc = window.bbox().loc + window_location;
+
+        if !outputs.iter().any(|o_geo| o_geo.contains(geo_loc)) {
+            orphaned_windows.push(window.clone());
+        }
+    }
+    for window in orphaned_windows.into_iter() {
+        place_new_window(space, &window, false);
+    }
+}
